refactor: #102 -  validation 구조 공통화 및 schema 분리
