+++ src/features/auth/signup/mutations/signupMutation.ts
@@ -1,5 +1,5 @@
 import { buildSignupRequestPayload } from "../lib/buildSignupRequestPayload";
-import { signupSuccessResponseSchema } from "../schema/schema";
+import { signupSuccessResponseSchema } from "../schema/signupSuccessReponseSchema";
 
 // 회원가입 요청 시 사용하는 payload 타입
 type SignupPayload = {
